Last improvements (description below)  - better Program{1-4} representation, migration  - small Movement rework (due to Program rework)  - commenting  - head/back touch working, head/back align  - removed Point (with actual logic it's useless)  - bug fixes

+++ src/cz/matejkripner/core/Measurement.java
@@ -6,27 +6,37 @@ import java.util.HashMap;
 import java.util.Map;
 
 /**
+ * Measurement base
  * @author Matěj Kripner <kripnermatej@gmail.com>
- * @version 1.0
+ * @author Jakub Vaněk <vanek.jakub4@seznam.cz>
+ * @version 1.1
  */
 public enum Measurement {
-
+	/**
+	 * Back-crash measurement
+	 */
     BACK_TOUCH {
         @Override
         protected Object doMeasurement(Hardware hardware) {
             return hardware.backTouch();
         }
     },
+	/**
+	 * Head-crash measurement
+	 */
     HEAD_TOUCH {
         @Override
         protected Object doMeasurement(Hardware hardware) {
             return hardware.headTouch();
         }
     },
+	/**
+	 * Ultrasonic sensor measurement
+	 */
     ULTRASONIC {
         @Override
         protected Object doMeasurement(Hardware hardware) {
-            Map<Hardware.Direction, Integer> measurements = new HashMap<>();
+            Map<Hardware.Direction, Float> measurements = new HashMap<>();
             measurements.put(hardware.getSonicDirection(), hardware.sonic());
             Hardware.Direction[] process = hardware.getSonicDirection().processAll();
             hardware.turnSonic(process[0]);
@@ -36,20 +46,35 @@ public enum Measurement {
             return new SonicMeasurement(measurements);
         }
     },
+	/**
+	 * Gyroscope measurement
+	 */
     GYRO {
         @Override
         protected Object doMeasurement(Hardware hardware) {
             return hardware.gyro();
         }
     };
-
+	/**
+	 * Whether is this measurement actual or not
+	 */
     private boolean actual;
+	/**
+	 * Cache of this measurement
+	 */
     private Object value;
 
+	/**
+	 * Default constructor
+	 */
     Measurement() {
         actual = true;
     }
 
+	/**
+	 * Get measurement
+	 * @return Cached (or updated) measurement
+	 */
     public Object get() {
         if(isActual()) return value;
         else {
@@ -58,20 +83,36 @@ public enum Measurement {
         }
     }
 
+	/**
+	 * Check whether this measurement is actual
+	 * @return True if actual
+	 */
     public boolean isActual() {
         return actual;
     }
 
+	/**
+	 * Set the actuality status of this measurement
+	 * @param actual Status
+	 */
     public void setActual(boolean actual) {
         this.actual = actual;
     }
 
+	/**
+	 * Mark all measurements outdated (not actual)
+	 */
     public static void allOutdated() {
         for(Measurement m : Measurement.values()) {
             m.setActual(false);
         }
     }
 
+	/**
+	 * Measure this physical quantity
+	 * @param hardware Robot's hardware
+	 * @return Measurement
+	 */
     protected Object doMeasurement(Hardware hardware) {
         throw new AbstractMethodError();
     }
